# memoriam/data/openapi_spec_storage.yml
openapi: 3.1.0
info:
  title: Memoriam Storage API
  description: OpenAPI spec for object storage
  version: '1.0'
  x-logo:
    url: /docs/_static/memoriam.png
    altText: Memoriam
servers:
  - url: /api
...
paths:
  /storage/buckets:
    get:
      operationId: memoriam.storage.storage.list_buckets
      summary: List buckets
      tags: [Storage]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    bucket_name:
                      type: string
                    creation_date:
                      type: string
                      format: date-time
        '401':
          description: Unauthorized
        '404':
          description: Not found
    post:
      operationId: memoriam.storage.storage.post_bucket
      summary: Create a new bucket
      tags: [Storage]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                bucket_name: *bucket_name_schema
              required:
                - bucket_name
      responses:
        '201':
          description: Created
        '400':
          description: Bad input parameters
        '401':
          description: Unauthorized
        '409':
          description: Conflict (name already exists)
...